requirements.txt
python-dotenv
langchain
langchain_openai
langchain_community
sqlalchemy
pandas
beautifulsoup4
langgraph
faiss-cpu
selenium